Made better use of debug levels with FCurve. Made better use of debug levels with FileSequence.
git-svn-id: svn://svn.blur.com/blurdev/trunk/code/python/lib/blur3d/api@19325 896c8824-d3f5-0310-9c61-ac502b9a79dc

+++ cross3d/classes/filesequence.py
@@ -12,6 +12,7 @@
 
 import os
 import re
+import blurdev
 import subprocess
 
 from framerange import FrameRange
@@ -264,9 +265,15 @@ class FileSequence( object ):
 			elif videoCodec == VideoCodec.H264:
 				command = [ffmpeg ]
 
-			process = subprocess.Popen( command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.PIPE )
+			if blurdev.debug.debugLevel() >= blurdev.debug.DebugLevel.Mid: 
+				print 'SEQUENCE TO MOVIE COMMAND: {}'.format(' '.join(command))
+
+			process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, stdin=subprocess.PIPE)
 			process.communicate()
-			normalisedSequence.delete()
+
+			if blurdev.debug.debugLevel() < blurdev.debug.DebugLevel.Mid: 
+				normalisedSequence.delete()
+				
 			return True
 		else:
 			raise Exception( 'Input sequence %s is missing frames' % self._path )
